+++ src/sections/traits/elevation.js
@@ -8,6 +8,12 @@ export default {
 	 *
 	 * @param {number|[number, number]} inElevation
 	 * @param {Object} inOptions
+	 * @param {boolean} [inOptions.absolute=false]
+	 * @param {boolean} [inOptions.topInclusive=false] Range form only.
+	 *   When true, the top value is part of the range, matching Foundry
+	 *   Region `elevation.topInclusive`. By default an effect whose top
+	 *   equals the next level's bottom sits just below that level rather
+	 *   than reaching into it.
 	 * @returns {this}
 	 */
 	elevation(inElevation, inOptions = {}) {
@@ -38,6 +44,7 @@ export default {
 			{
 				elevation: 1,
 				absolute: false,
+				topInclusive: false,
 			},
 			inOptions
 		);
@@ -48,10 +55,17 @@ export default {
 				"elevation",
 				"inOptions.absolute must be of type boolean"
 			);
+		if (typeof inOptions.topInclusive !== "boolean")
+			throw this.sequence._customError(
+				this,
+				"elevation",
+				"inOptions.topInclusive must be of type boolean"
+			);
 		this._elevation = {
 			elevation: bottom,
 			top,
 			absolute: inOptions.absolute,
+			topInclusive: inOptions.topInclusive,
 		};
 		return this;
 	},